split banks component
split banks component
