+++ src/client/components/banks/banks.jsx
@@ -1,52 +1,51 @@
 import React from "react";
-import Slider from "react-slick";
 import ReadMoreText from "../read-more-text/read-more-text";
+import MobileBanks from "../banks-mobile/banks-mobile";
 
-import "./banks.css";
 import logo1 from "./Blackstone.jpg";
 import logo2 from "./Palazzo.png";
 import logo3 from "./JPMorgan Chase & Co..png";
+import "./banks.css";
 
-const Banks = () => {
-  const settings = {
-    dots: true,
-    infinite: false,
-    speed: 500,
-    slidesToShow: 1,
-    slidesToScroll: 1,
-  };
+const Banks = () => (
+  <section id="banks-section">
+    <div className="mobile-title">
+      <h3 className="what-does-bankeradvisor-offer-mobile">
+        What Does BankerAdvisor Offer?
+      </h3>
+    </div>
 
-  return (
-    <section id="banks-section">
-      <div className="mobile-title">
-        <h3 className="what-does-bankeradvisor-offer-mobile">
+    <MobileBanks />
+
+    <div className="banks">
+      <div className="banks-wrap">
+        <h3 className="what-does-bankeradvisor-offer-header">
           What Does BankerAdvisor Offer?
         </h3>
-      </div>
-      <div className="mobile-bank-wrap">
-        <Slider {...settings}>
-          <div className="mobile-bank">
-            <div className="mobile-bank-header">
-              <div className="mobile-logo">
-                <img className="mobile-logo-img" src={logo1} alt="" />
-              </div>
-              <div className="mobile-link-wrapp">
-                <a className="mobile-link" href="https://www.blackstone.com/">
+        <div className="bank">
+          <div className="bank-logo">
+            <img className="bank-logo-img" src={logo1} alt="bank logo" />
+          </div>
+          <div className="bank-description">
+            <div className="bank-name">
+              <h4 className="bank-name-header">Blackstone</h4>
+              <div className="bank-name-website-link-wrapp">
+                <a
+                  className="bank-name-website-link"
+                  href="https://www.blackstone.com/"
+                >
                   blackstone.com
                 </a>
               </div>
             </div>
-            <div className="mobile-title">
-              <h4>Blackstone</h4>
-            </div>
-            <div className="mobile-industry">
-              <p className="mobile-industry-services">
-                Private Equity, Real Estate Infrastructure, Innovations, Life
-                Sciences
-              </p>
-              <p className="mobile-industry-employees">1,000-5,000 employees</p>
-            </div>
-            <div className="mobile-text">
+            <div className="bank-industry-wrapp">
+              <div className="bank-industry">
+                <p className="bank-industry-services">
+                  Private Equity, Real Estate Infrastructure, Innovations, Life
+                  Sciences
+                </p>
+                <p className="bank-industry-employees">1,000-5,000 employees</p>
+              </div>
               <ReadMoreText
                 text="Blackstone is one of the world’s leading investment firms.
                   Asset management businesses, with $450 billion in assets under
@@ -59,69 +58,73 @@ const Banks = () => {
                   services."
               />
             </div>
-            {/* <small>Read more</small> */}
-
             <div className="border-bottom" />
-
-            <div className="connect-mail">
-              <div className="connect-name">
-                <h5 className="connect-name-header">
-                  Alexandra Hill, Sr. Managing Director
-                </h5>
+            <div className="connect-person">
+              <div className="connect-person-item">
+                <div className="connect-name-email-wrapp">
+                  <p className="connect-person-name">
+                    Alexandra Hill, Sr. Managing Director
+                  </p>
+                  <a
+                    className="connect-person-email"
+                    href="mailto:alexandra.hill@blackstone.com"
+                  >
+                    alexandra.hill@blackstone.com
+                  </a>
+                </div>
                 <a
-                  className="connect-name-email-link"
+                  className="connect-person-link"
                   href="mailto:alexandra.hill@blackstone.com"
                 >
                   connect
                 </a>
               </div>
-              <a
-                className="connect-name-emali"
-                href="mailto:alexandra.hill@blackstone.com"
-              >
-                alexandra.hill@blackstone.com
-              </a>
-            </div>
-            <div className="connect-mail">
-              <div className="connect-name">
-                <h5 className="connect-name-header">
-                  Eli Nagler, Managing Director
-                </h5>
+              <div className="connect-person-item">
+                <div className="connect-name-email-wrapp">
+                  <p className="connect-person-name">
+                    Eli Nagler, Managing Director
+                  </p>
+                  <a
+                    className="connect-person-email"
+                    href="mailto:eli.nagler@blackstone.com"
+                  >
+                    eli.nagler@blackstone.com
+                  </a>
+                </div>
                 <a
-                  className="connect-name-email-link"
+                  className="connect-person-link"
                   href="mailto:eli.nagler@blackstone.com"
                 >
                   connect
                 </a>
               </div>
-              <a
-                className="connect-name-email"
-                href="mailto:eli.nagler@blackstone.com"
-              >
-                eli.nagler@blackstone.com
-              </a>
             </div>
           </div>
-          <div className="mobile-bank">
-            <div className="mobile-bank-header">
-              <div className="mobile-logo">
-                <img className="mobile-logo-img" src={logo2} alt="" />
+        </div>
+        <div className="bank">
+          <div className="bank-logo">
+            <img className="bank-logo-img" src={logo2} alt="bank logo" />
+          </div>
+          <div className="bank-description">
+            <div className="bank-name">
+              <h4 className="bank-name-header">Palazzo</h4>
+              <div className="bank-name-website-link-wrapp">
+                <a
+                  className="bank-name-website-link"
+                  href="https://www.palazzonyc.com/"
+                >
+                  palazzonyc.com
+                </a>
               </div>
-              <div className="mobile-link">
-                <a href="http://palazzonyc.com/">palazzonyc.com</a>
+            </div>
+            <div className="bank-industry-wrapp">
+              <div className="bank-industry">
+                <p className="bank-industry-services">
+                  Marketing, Media, Tech Infrastructure, Innovations, Life
+                  Sciences
+                </p>
+                <p className="bank-industry-employees">11-50 employees</p>
               </div>
-            </div>
-            <div className="mobile-title">
-              <h4>Palazzo</h4>
-            </div>
-            <div className="mobile-industry">
-              <p className="mobile-industry-services">
-                Marketing, Media, Tech Infrastructure, Innovations, Life
-                Sciences
-              </p>
-              <p className="mobile-industry-employees">11-50 employees</p>
-            </div>
-            <div className="mobile-text">
               <ReadMoreText
                 text="PALAZZO is a New York City based investment bank that provides M&A advisory
                   and capital raising services to entrepreneurs/founders, private equity firms and strategic acquirers
@@ -134,72 +137,70 @@ const Banks = () => {
                   in the sectors it covers."
               />
             </div>
-            {/* <small>Read more</small> */}
-
             <div className="border-bottom" />
-
-            <div className="connect-mail">
-              <div className="connect-name">
-                <h5 className="connect-name-header">
-                  Peter Cosco, Managing Director
-                </h5>
+            <div className="connect-person">
+              <div className="connect-person-item">
+                <div className="connect-name-email-wrapp">
+                  <p className="connect-person-name">
+                    Peter Cosco, Managing Director
+                  </p>
+                  <a
+                    className="connect-person-email"
+                    href="mailto:pcosco@palazzonyc.com"
+                  >
+                    pcosco@palazzonyc.com
+                  </a>
+                </div>
                 <a
-                  className="connect-name-email-link"
+                  className="connect-person-link"
                   href="mailto:pcosco@palazzonyc.com"
                 >
                   connect
                 </a>
               </div>
-              <a
-                className="connect-name-email"
-                href="mailto:pcosco@palazzonyc.com"
-              >
-                pcosco@palazzonyc.com
-              </a>
-            </div>
-            <div className="connect-mail">
-              <div className="connect-name">
-                <h5 className="connect-name-header">
-                  Matt Hasson, Managing Director
-                  {" "}
-                </h5>
+              <div className="connect-person-item">
+                <div className="connect-name-email-wrapp">
+                  <p className="connect-person-name">
+                    Matt Hasson, Managing Director
+                  </p>
+                  <a
+                    className="connect-person-email"
+                    href="mailto:mhasson@palazzonyc.com"
+                  >
+                    mhasson@palazzonyc.com
+                  </a>
+                </div>
                 <a
-                  className="connect-name-email-link"
+                  className="connect-person-link"
                   href="mailto:mhasson@palazzonyc.com"
                 >
                   connect
                 </a>
               </div>
-              <a
-                className="connect-name-email"
-                href="mailto:mhasson@palazzonyc.com"
-              >
-                mhasson@palazzonyc.com
-              </a>
             </div>
           </div>
-          <div className="mobile-bank">
-            <div className="mobile-bank-header">
-              <div className="mobile-logo">
-                <img className="mobile-logo-img" src={logo3} alt="" />
-              </div>
-              <div className="mobile-link-wrapp">
+        </div>
+        <div className="bank">
+          <div className="bank-logo">
+            <img className="bank-logo-img" src={logo3} alt="bank logo" />
+          </div>
+          <div className="bank-description">
+            <div className="bank-name">
+              <h4 className="bank-name-header">JPMorgan Chase & Co.</h4>
+              <div className="bank-name-website-link-wrapp">
                 <a
-                  className="mobile-link"
+                  className="bank-name-website-link"
                   href="https://www.jpmorganchase.com/"
                 >
                   jpmorganchase.com
                 </a>
               </div>
             </div>
-            <div className="mobile-title">
-              <h4>JPMorgan Chase & Co.</h4>
-            </div>
-            <div className="mobile-industry">
-              <p className="mobile-industry-services">Financial Services</p>
-              <p className="mobile-industry-employees">10,000+ employees</p>
-            </div>
-            <div className="mobile-text">
+            <div className="bank-industry-wrapp">
+              <div className="bank-industry">
+                <p className="bank-industry-services">Financial Services </p>
+                <p className="bank-industry-employees">10,000+ employees</p>
+              </div>
               <ReadMoreText
                 text="For over 200 years, JPMorgan Chase & Co. has provided innovative financial
                   solutions for consumers, small businesses, corporations,
@@ -210,296 +211,52 @@ const Banks = () => {
                   the future and strengthen both our clients and our communities."
               />
             </div>
-            {/* <small>Read more</small> */}
-
             <div className="border-bottom" />
-
-            <div className="connect-mail">
-              <div className="connect-name">
-                <h5 className="connect-name-header">
-                  David Disi, Chief Investment Officer
-                </h5>
+            <div className="connect-person">
+              <div className="connect-person-item">
+                <div className="connect-name-email-wrapp">
+                  <p className="connect-person-name">
+                    David Disi, Chief Investment Officer
+                  </p>
+                  <a
+                    className="connect-person-email"
+                    href="mailto:david@jpmorgan.com"
+                  >
+                    david@jpmorgan.com
+                  </a>
+                </div>
                 <a
-                  className="connect-name-email-link"
+                  className="connect-person-link"
                   href="mailto:david@jpmorgan.com"
                 >
                   connect
                 </a>
               </div>
-              <a
-                className="connect-name-email"
-                href="mailto:david@jpmorgan.com"
-              >
-                david@jpmorgan.com
-                {" "}
-              </a>
-            </div>
-            <div className="connect-mail">
-              <div className="connect-name">
-                <h5 className="connect-name-header">
-                  Andrea Stefanucci, Executive Director
-                </h5>
+              <div className="connect-person-item">
+                <div className="connect-name-email-wrapp">
+                  <p className="connect-person-name">
+                    Andrea Stefanucci, Executive Director
+                  </p>
+                  <a
+                    className="connect-person-email"
+                    href="mailto:astefanucci@jpmorgan.com"
+                  >
+                    astefanucci@jpmorgan.com
+                  </a>
+                </div>
                 <a
-                  className="connect-name-email-link"
+                  className="connect-person-link"
                   href="mailto:astefanucci@jpmorgan.com"
                 >
                   connect
                 </a>
               </div>
-              <a
-                className="connect-name-email"
-                href="mailto:astefanucci@jpmorgan.com"
-              >
-                astefanucci@jpmorgan.com
-              </a>
-            </div>
-          </div>
-        </Slider>
-      </div>
-
-      <div className="banks">
-        <div className="banks-wrap">
-          <h3 className="what-does-bankeradvisor-offer-header">
-            What Does BankerAdvisor Offer?
-          </h3>
-          <div className="bank">
-            <div className="bank-logo">
-              <img className="bank-logo-img" src={logo1} alt="bank logo" />
-            </div>
-            <div className="bank-description">
-              <div className="bank-name">
-                <h4 className="bank-name-header">Blackstone</h4>
-                <div className="bank-name-website-link-wrapp">
-                  <a
-                    className="bank-name-website-link"
-                    href="https://www.blackstone.com/"
-                  >
-                    blackstone.com
-                  </a>
-                </div>
-              </div>
-              <div className="bank-industry-wrapp">
-                <div className="bank-industry">
-                  <p className="bank-industry-services">
-                    Private Equity, Real Estate Infrastructure, Innovations,
-                    Life Sciences
-                  </p>
-                  <p className="bank-industry-employees">
-                    1,000-5,000 employees
-                  </p>
-                </div>
-                <ReadMoreText
-                  text="Blackstone is one of the world’s leading investment firms.
-                  Asset management businesses, with $450 billion in assets under
-                  management, include investment vehicles focused on private
-                  equity, real estate, public debt and equity, non-investment
-                  grade credit, real assets and secondary funds, all on a global
-                  basis. Blackstone also provides various financial advisory
-                  services, including financial and strategic advisory,
-                  restructuring and reorganization advisory and fund placement
-                  services."
-                />
-              </div>
-              <div className="border-bottom" />
-              <div className="connect-person">
-                <div className="connect-person-item">
-                  <div className="connect-name-email-wrapp">
-                    <p className="connect-person-name">
-                      Alexandra Hill, Sr. Managing Director
-                    </p>
-                    <a
-                      className="connect-person-email"
-                      href="mailto:alexandra.hill@blackstone.com"
-                    >
-                      alexandra.hill@blackstone.com
-                    </a>
-                  </div>
-                  <a
-                    className="connect-person-link"
-                    href="mailto:alexandra.hill@blackstone.com"
-                  >
-                    connect
-                  </a>
-                </div>
-                <div className="connect-person-item">
-                  <div className="connect-name-email-wrapp">
-                    <p className="connect-person-name">
-                      Eli Nagler, Managing Director
-                    </p>
-                    <a
-                      className="connect-person-email"
-                      href="mailto:eli.nagler@blackstone.com"
-                    >
-                      eli.nagler@blackstone.com
-                    </a>
-                  </div>
-                  <a
-                    className="connect-person-link"
-                    href="mailto:eli.nagler@blackstone.com"
-                  >
-                    connect
-                  </a>
-                </div>
-              </div>
-            </div>
-          </div>
-          <div className="bank">
-            <div className="bank-logo">
-              <img className="bank-logo-img" src={logo2} alt="bank logo" />
-            </div>
-            <div className="bank-description">
-              <div className="bank-name">
-                <h4 className="bank-name-header">Palazzo</h4>
-                <div className="bank-name-website-link-wrapp">
-                  <a
-                    className="bank-name-website-link"
-                    href="https://www.palazzonyc.com/"
-                  >
-                    palazzonyc.com
-                  </a>
-                </div>
-              </div>
-              <div className="bank-industry-wrapp">
-                <div className="bank-industry">
-                  <p className="bank-industry-services">
-                    Marketing, Media, Tech Infrastructure, Innovations, Life
-                    Sciences
-                  </p>
-                  <p className="bank-industry-employees">11-50 employees</p>
-                </div>
-                <ReadMoreText
-                  text="PALAZZO is a New York City based investment bank that provides M&A advisory
-                  and capital raising services to entrepreneurs/founders, private equity firms and strategic acquirers
-                  in the marketing services, media, and technology sectors.
-                  The firm is unique in its ability to offer a powerful combination
-                  of investment banking and advisory services,
-                  all designed to maximize value and wealth for its clients,
-                  no matter where they are in their business lifecycle.
-                  Founded in 2009, the firm has quickly grown into a leading investment bank
-                  in the sectors it covers."
-                />
-              </div>
-              <div className="border-bottom" />
-              <div className="connect-person">
-                <div className="connect-person-item">
-                  <div className="connect-name-email-wrapp">
-                    <p className="connect-person-name">
-                      Peter Cosco, Managing Director
-                    </p>
-                    <a
-                      className="connect-person-email"
-                      href="mailto:pcosco@palazzonyc.com"
-                    >
-                      pcosco@palazzonyc.com
-                    </a>
-                  </div>
-                  <a
-                    className="connect-person-link"
-                    href="mailto:pcosco@palazzonyc.com"
-                  >
-                    connect
-                  </a>
-                </div>
-                <div className="connect-person-item">
-                  <div className="connect-name-email-wrapp">
-                    <p className="connect-person-name">
-                      Matt Hasson, Managing Director
-                    </p>
-                    <a
-                      className="connect-person-email"
-                      href="mailto:mhasson@palazzonyc.com"
-                    >
-                      mhasson@palazzonyc.com
-                    </a>
-                  </div>
-                  <a
-                    className="connect-person-link"
-                    href="mailto:mhasson@palazzonyc.com"
-                  >
-                    connect
-                  </a>
-                </div>
-              </div>
-            </div>
-          </div>
-          <div className="bank">
-            <div className="bank-logo">
-              <img className="bank-logo-img" src={logo3} alt="bank logo" />
-            </div>
-            <div className="bank-description">
-              <div className="bank-name">
-                <h4 className="bank-name-header">JPMorgan Chase & Co.</h4>
-                <div className="bank-name-website-link-wrapp">
-                  <a
-                    className="bank-name-website-link"
-                    href="https://www.jpmorganchase.com/"
-                  >
-                    jpmorganchase.com
-                  </a>
-                </div>
-              </div>
-              <div className="bank-industry-wrapp">
-                <div className="bank-industry">
-                  <p className="bank-industry-services">Financial Services </p>
-                  <p className="bank-industry-employees">10,000+ employees</p>
-                </div>
-                <ReadMoreText
-                  text="For over 200 years, JPMorgan Chase & Co. has provided innovative financial
-                  solutions for consumers, small businesses, corporations,
-                  governments and institutions around the world.
-                  Today, we're a leading global financial services firm with operations servicing clients in
-                  more than 100 countries. Whether we are serving customers, helping small businesses, or putting our skills to work with partners,
-                  we strive to identify issues and propose solutions that will propel
-                  the future and strengthen both our clients and our communities."
-                />
-              </div>
-              <div className="border-bottom" />
-              <div className="connect-person">
-                <div className="connect-person-item">
-                  <div className="connect-name-email-wrapp">
-                    <p className="connect-person-name">
-                      David Disi, Chief Investment Officer
-                    </p>
-                    <a
-                      className="connect-person-email"
-                      href="mailto:david@jpmorgan.com"
-                    >
-                      david@jpmorgan.com
-                    </a>
-                  </div>
-                  <a
-                    className="connect-person-link"
-                    href="mailto:david@jpmorgan.com"
-                  >
-                    connect
-                  </a>
-                </div>
-                <div className="connect-person-item">
-                  <div className="connect-name-email-wrapp">
-                    <p className="connect-person-name">
-                      Andrea Stefanucci, Executive Director
-                    </p>
-                    <a
-                      className="connect-person-email"
-                      href="mailto:astefanucci@jpmorgan.com"
-                    >
-                      astefanucci@jpmorgan.com
-                    </a>
-                  </div>
-                  <a
-                    className="connect-person-link"
-                    href="mailto:astefanucci@jpmorgan.com"
-                  >
-                    connect
-                  </a>
-                </div>
-              </div>
             </div>
           </div>
         </div>
       </div>
-    </section>
-  );
-};
+    </div>
+  </section>
+);
 
 export default Banks;
