+++ src/client/components/get-banks/get-banks.css
@@ -93,7 +93,7 @@
   .share-banks {
     padding: 0;
     margin-top: -50px;
-    margin-left: -20px;
+    margin-left: 20px;
     width: 450px;
   }
   .share-banks p:first-child {
